active_contours/requirements.txt
opencv-python~=4.1.2.30
numpy~=1.19.5
Pillow~=8.1.2
PyQt5~=5.15.4